Refactor backlink component
Refactor backlink and renamed component to make it work with the new code structure.
Adjusted styling for desktop/mobile view.
