--- src/app/components/back-link/backLink.less
@@ -1,19 +1,0 @@
-@import '../../../../node_modules/nav-frontend-lenker-style/src/lenker-style';
-@import '../../../../node_modules/nav-frontend-core/less/_variabler';
-
-.backLink {
-	color: @navMorkGra;
-	text-decoration: none;
-	font-weight: bold;
-	display: flex;
-	align-items: center;
-
-	&:hover,
-	&:active {
-		color: @navBla;
-	}
-}
-
-.lenke {
-	border-bottom: none;
-}
