--- src/app/components/progress-indicator/stepIndicator.less
@@ -1,5 +1,0 @@
-.stegindikatortittel {
-	text-align: center;
-	margin: 0;
-	padding-bottom: 12px;
-}
