+++ src/app/components/step-indicator/Step.tsx
