+++ src/app/components/step-indicator/StepIndicator.tsx
