+++ src/app/components/step-indicator/stepIndicator.less
@@ -1,0 +1,17 @@
+@import '../../styles/variables.less';
+.stegindikatortittel {
+    text-align: center;
+    margin: 0;
+    padding-bottom: 12px;
+    .typo-ingress {
+        margin-top: @marginMedium !important;
+    }
+}
+
+@media screen and (min-width: @screen-desktop) {
+    .stegindikatortittel {
+        .typo-ingress {
+            margin-top: @marginLarge !important;
+        }
+    }
+}
